+++ core/src/cats/cats_backends.cc
@@ -2,7 +2,7 @@
    BAREOS® - Backup Archiving REcovery Open Sourced
 
    Copyright (C) 2011-2016 Planets Communications B.V.
-   Copyright (C) 2013-2021 Bareos GmbH & Co. KG
+   Copyright (C) 2013-2023 Bareos GmbH & Co. KG
 
    This program is Free Software; you can redistribute it and/or
    modify it under the terms of version three of the GNU Affero General Public
@@ -39,11 +39,11 @@ static struct backend_interface_mapping_t {
   bool partly_compare;
   int interface_type_id;
 } backend_interface_mappings[]
-    = {{"dbi", TRUE, SQL_INTERFACE_TYPE_DBI},
-       {"mysql", FALSE, SQL_INTERFACE_TYPE_MYSQL},
-       {"postgresql", FALSE, SQL_INTERFACE_TYPE_POSTGRESQL},
-       {"sqlite3", FALSE, SQL_INTERFACE_TYPE_SQLITE3},
-       {NULL, FALSE, 0}};
+    = {{"dbi", true, SQL_INTERFACE_TYPE_DBI},
+       {"mysql", false, SQL_INTERFACE_TYPE_MYSQL},
+       {"postgresql", false, SQL_INTERFACE_TYPE_POSTGRESQL},
+       {"sqlite3", false, SQL_INTERFACE_TYPE_SQLITE3},
+       {nullptr, false, 0}};
 
 #    include "cats_backends.h"
 #    include <dlfcn.h>
@@ -174,10 +174,8 @@ BareosDb* db_init_database(JobControlRecord* jcr,
          _("Driver type not specified in Catalog resource.\n"));
   }
 
-  /*
-   * If we didn't find a mapping its fatal because we don't know what database
-   * backend to use.
-   */
+  /* If we didn't find a mapping its fatal because we don't know what database
+   * backend to use. */
   backend_interface_mapping = lookup_backend_interface_mapping(db_driver);
   if (backend_interface_mapping == NULL) {
     Jmsg(jcr, M_ERROR_TERM, 0, _("Unknown database driver: %s\n"), db_driver);
@@ -197,10 +195,8 @@ BareosDb* db_init_database(JobControlRecord* jcr,
     }
   }
 
-  /*
-   * This is a new backend try to use dynamic loading to load the backend
-   * library.
-   */
+  /* This is a new backend try to use dynamic loading to load the backend
+   * library. */
 
 #    if defined HAVE_WIN32
   Mmsg(shared_library_name, "libbareoscats-%s%s",
@@ -231,10 +227,8 @@ BareosDb* db_init_database(JobControlRecord* jcr,
 #    endif
 
   if (backend_shared_library) {
-    /*
-     * Create a new loaded shared library entry and tack it onto the list of
-     * loaded backend shared libs.
-     */
+    /* Create a new loaded shared library entry and tack it onto the list of
+     * loaded backend shared libs. */
 
     if (loaded_backends == NULL) {
       loaded_backends